various sundry game object refactors
