+++ TODO
@@ -1,3 +1,5 @@
+Plating:
+  ☐ link to source on game page
 Refactors:
   ☐ Items To Config Files
 Additional Features:
