+++ aspirant_application/Common.XY.h
@@ -37,9 +37,4 @@ namespace common
 	{
 		return first.GetX() == second.GetX() && first.GetY() == second.GetY();
 	}
-	template<typename TDimension>
-	bool operator<(const XY<TDimension>& first, const XY<TDimension>& second)
-	{
-		return first.GetY() < second.GetY() || first.GetX() < second.GetX();
-	}
 }
