+++ aspirant_application/Game.World.Border.h
@@ -1,0 +1,9 @@
+#pragma once
+namespace game::world
+{
+	enum class Border
+	{
+		WALL,
+		DOOR
+	};
+}
