+++ aspirant_application/Game.World.h
@@ -1,19 +1,15 @@
 #pragma once
 #include "Common.XY.h"
 #include "Maze.Direction.h"
+#include "Game.World.Border.h"
 namespace game::World
 {
-	enum class Border
-	{
-		WALL,
-		DOOR
-	};
-	Border GetLeftSide();
-	Border GetAhead();
-	Border GetRightSide();
+	world::Border GetLeftSide();
+	world::Border GetAhead();
+	world::Border GetRightSide();
 
 	common::XY<size_t> GetSize();
-	Border GetBorder(const common::XY<size_t>&, const maze::Direction&);
+	world::Border GetBorder(const common::XY<size_t>&, const maze::Direction&);
 	bool IsExplored(const common::XY<size_t>&);
 	void SetExplored(const common::XY<size_t>&);
 }
