+++ aspirant_application/Game.cpp
@@ -2,7 +2,7 @@
 namespace game::World { void Reset(); }
 namespace game::Avatar { void Reset(); }
 namespace game::avatar::Statistics { void Reset(); }
-namespace game::Items { void Reset(); }
+namespace game::world::Items { void Reset(); }
 namespace game::avatar::Items { void Reset(); }
 namespace game::Creatures { void Reset(); }
 namespace game
@@ -10,7 +10,7 @@ namespace game
 	void Reset()
 	{
 		game::World::Reset();
-		game::Items::Reset();
+		game::world::Items::Reset();
 		game::avatar::Statistics::Reset();
 		game::avatar::Items::Reset();
 		game::Creatures::Reset();
