+++ aspirant_application/aspirant_application.vcxproj
@@ -177,6 +177,7 @@
     <ClCompile Include="Game.Creatures.cpp" />
     <ClCompile Include="Game.Items.cpp" />
     <ClCompile Include="Game.World.cpp" />
+    <ClCompile Include="Game.World.Items.cpp" />
     <ClCompile Include="Graphics.AvatarInventory.cpp" />
     <ClCompile Include="Graphics.FloorInventory.cpp" />
     <ClCompile Include="Graphics.Images.cpp" />
@@ -252,7 +253,9 @@
     <ClInclude Include="Game.h" />
     <ClInclude Include="Game.Item.h" />
     <ClInclude Include="Game.Items.h" />
+    <ClInclude Include="Game.World.Border.h" />
     <ClInclude Include="Game.World.h" />
+    <ClInclude Include="Game.World.Items.h" />
     <ClInclude Include="Graphics.AvatarInventory.h" />
     <ClInclude Include="Graphics.FloorInventory.h" />
     <ClInclude Include="Graphics.Images.h" />
