+++ jam.TODO
@@ -1,6 +1,10 @@
 Blocked:
     ☐ Welcome to The Hive! https://itch.io/jam/dcjam2021/rate/985851
         need "steam friends"
+    ☐ Calderan Chronicles - Prologue https://itch.io/jam/dcjam2021/rate/987105
+        stream killer
+    ☐ Sunken https://itch.io/jam/dcjam2021/rate/987313
+        win7
 20210408:
     ✔ Blob Keep https://itch.io/jam/dcjam2021/rate/978972 @done(21-04-08 04:10)
     needs a greater time investment
@@ -20,25 +24,23 @@ Blocked:
     ✔ Space Crawlers #dcjam2021 https://itch.io/jam/dcjam2021/rate/986649 @done(21-04-08 05:51)
     ✔ The Infinite Grimoire https://itch.io/jam/dcjam2021/rate/986377 @done(21-04-08 05:58)
         need winrar
-20210409(planned... we'll see how it actually goes):
-    ☐ The Reign of the Lizard Wizard https://itch.io/jam/dcjam2021/rate/986676
+20210409:
+    ✔ The Reign of the Lizard Wizard https://itch.io/jam/dcjam2021/rate/986676 @done(21-04-09 04:09)
         need winrar
-    ☐ Mystidinia https://itch.io/jam/dcjam2021/rate/986114
-    ☐ Calderan Chronicles - Prologue https://itch.io/jam/dcjam2021/rate/987105
-    ☐ The Dark Forest in Heaven https://itch.io/jam/dcjam2021/rate/987111
-    ☐ Synthyfriends Dungeon https://itch.io/jam/dcjam2021/rate/987118
-    ☐ Tria Insomnia https://itch.io/jam/dcjam2021/rate/986299
-    ☐ Sunken https://itch.io/jam/dcjam2021/rate/987313
-    ☐ Terminus-13 https://itch.io/jam/dcjam2021/rate/987192
-    ☐ Pages of Arkham https://itch.io/jam/dcjam2021/rate/986117
-    ☐ Goblin Slaying Dungeon Crawler https://itch.io/jam/dcjam2021/rate/987400
-    ☐ Two Problems: A Regular Space Adventure https://itch.io/jam/dcjam2021/rate/987164
-    ☐ This Mall is a Maze and I Gotta Pee https://itch.io/jam/dcjam2021/rate/986502
-Jam Games To Review:
-    ☐ Hotel Inverse https://itch.io/jam/dcjam2021/rate/986943
-    ☐ Prototype dungeon crawler https://itch.io/jam/dcjam2021/rate/987495
-    ☐ Ankehotep's Prison https://itch.io/jam/dcjam2021/rate/987063
-    ☐ Psycho Princess: The Labyrinth https://itch.io/jam/dcjam2021/rate/987510
+    ✔ Mystidinia https://itch.io/jam/dcjam2021/rate/986114 @done(21-04-09 04:18)
+    ✔ The Dark Forest in Heaven https://itch.io/jam/dcjam2021/rate/987111 @done(21-04-09 04:33)
+    ✔ Synthyfriends Dungeon https://itch.io/jam/dcjam2021/rate/987118 @done(21-04-09 04:41)
+    ✔ Tria Insomnia https://itch.io/jam/dcjam2021/rate/986299 @done(21-04-09 04:46)
+    ✔ Terminus-13 https://itch.io/jam/dcjam2021/rate/987192 @done(21-04-09 04:53)
+    ✔ Pages of Arkham https://itch.io/jam/dcjam2021/rate/986117 @done(21-04-09 05:01)
+    ✔ Goblin Slaying Dungeon Crawler https://itch.io/jam/dcjam2021/rate/987400 @done(21-04-09 05:05)
+    ✔ Two Problems: A Regular Space Adventure https://itch.io/jam/dcjam2021/rate/987164 @done(21-04-09 05:10)
+    ✔ This Mall is a Maze and I Gotta Pee https://itch.io/jam/dcjam2021/rate/986502 @done(21-04-09 05:22)
+    ✔ Hotel Inverse https://itch.io/jam/dcjam2021/rate/986943 @done(21-04-09 05:28)
+    ✔ Prototype dungeon crawler https://itch.io/jam/dcjam2021/rate/987495 @done(21-04-09 05:32)
+    ✔ Ankehotep's Prison https://itch.io/jam/dcjam2021/rate/987063 @done(21-04-09 05:40)
+    ✔ Psycho Princess: The Labyrinth https://itch.io/jam/dcjam2021/rate/987510 @done(21-04-09 05:50)
+20210410:
     ☐ Space Station Escape! (Dungeon Crawler Jam 2021) https://itch.io/jam/dcjam2021/rate/987557
     ☐ MazeCrawler https://itch.io/jam/dcjam2021/rate/987562
     ☐ Characternik https://itch.io/jam/dcjam2021/rate/979498
@@ -51,6 +53,7 @@ Jam Games To Review:
     ☐ Mork's Trial https://itch.io/jam/dcjam2021/rate/987487
     ☐ The Dreadful Dungeon of Danger and Disaster https://itch.io/jam/dcjam2021/rate/987143
     ☐ Priest for the lich https://itch.io/jam/dcjam2021/rate/987446
+Jam Games To Review:
     ☐ Unfinished Dungeon https://itch.io/jam/dcjam2021/rate/987694
     ☐ Through the dungeon https://itch.io/jam/dcjam2021/rate/987710
     ☐ Ard'af https://itch.io/jam/dcjam2021/rate/978968
